Spotify Quick Control Implementation
Implementation of the field "Spotify_Control" as related to the proposed change discussed on actions.py

Reason for suggesting this: Google is rolling out new feature adjustments to this feature. In particular, if you own a chromecast or a chromecast-compilant device you can simply ask google to "put up some music on DEVICENAME" without any further configuration. But in order for this to work properly, Spotify implementation has to be disabled.